Move push notifications listeners to root component
